@@ -4,13 +4,17 @@ export declare class LeafletElement {
     private mapService;
     lat: number;
     lon: number;
+    x: number;
+    y: number;
     zoom: number;
     minZoom: number;
     maxZoom: number;
     layerControl: boolean;
+    crs: any;
     mapElement: ElementRef;
+    layerControlObject: any;
     constructor(mapService: MapService);
     ngOnInit(): void;
     ngAfterViewInit(): void;
-    loop(): void;
+    setLayerControl(): void;
 }
