@@ -11,73 +11,103 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 var __param = (this && this.__param) || function (paramIndex, decorator) {
     return function (target, key) { decorator(target, key, paramIndex); }
 };
-var core_1 = require('@angular/core');
-var map_1 = require('../map/map');
-var group_1 = require('../group/group');
-var map_service_1 = require('../services/map.service');
-var group_service_1 = require('../services/group.service');
-var popup_service_1 = require('../services/popup.service');
-var path_1 = require('../models/path');
+var core_1 = require("@angular/core");
+var map_1 = require("../map/map");
+var group_1 = require("../group/group");
+var map_service_1 = require("../services/map.service");
+var group_service_1 = require("../services/group.service");
+var popup_service_1 = require("../services/popup.service");
+var globalId_service_1 = require("../services/globalId.service");
+var helper_service_1 = require("../services/helper.service");
+var path_1 = require("../models/path");
 var Lealflet = require('leaflet');
 var PolylineElement = (function () {
-    function PolylineElement(mapService, groupService, popupService, LeafletElement, LeafletGroup) {
+    function PolylineElement(mapService, groupService, popupService, guidService, helperService, LeafletElement, LeafletGroup) {
         this.mapService = mapService;
         this.groupService = groupService;
         this.popupService = popupService;
+        this.guidService = guidService;
+        this.helperService = helperService;
         this.LeafletElement = LeafletElement;
         this.LeafletGroup = LeafletGroup;
         this.latlngs = [[52.6, -1.1], [52.605, -1.1], [52.606, -1.105], [52.697, -1.109]];
         this.Options = new path_1.path(null);
         this.mouseover = "";
         this.onclick = "";
+        this.polyline = null;
+        this.originalObject = this.latlngs.slice();
+        this.globalId = this.guidService.newGuid();
     }
     PolylineElement.prototype.ngOnInit = function () {
         if (this.LeafletElement || this.LeafletGroup) {
+            this.Options.fill = false;
             var inheritedOptions = new path_1.path(this.Options);
-            inheritedOptions.fill = false;
             var map = this.mapService.getMap();
-            var polyline = L.polyline(this.latlngs, inheritedOptions);
-            this.popupService.enablePopup(this.mouseover, this.onclick, polyline);
+            this.polyline = L.polyline(this.latlngs, inheritedOptions);
+            this.popupService.enablePopup(this.mouseover, this.onclick, this.polyline);
             if (this.LeafletGroup) {
-                this.groupService.addOLayersToGroup(polyline);
-                this.groupService.increaseNumber();
+                this.groupService.addOLayersToGroup(this.polyline, map, this.mapService, this.LeafletGroup, false, this.globalId);
             }
             else {
-                polyline.addTo(map);
+                this.polyline.addTo(map);
             }
         }
         else {
             console.warn("This polyline-element will not be rendered \n the expected parent node of polyline-element should be either leaf-element or leaflet-group");
         }
     };
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Array)
-    ], PolylineElement.prototype, "latlngs", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Object)
-    ], PolylineElement.prototype, "Options", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', String)
-    ], PolylineElement.prototype, "mouseover", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', String)
-    ], PolylineElement.prototype, "onclick", void 0);
-    PolylineElement = __decorate([
-        core_1.Component({
-            moduleId: module.id,
-            selector: 'polyline-element',
-            templateUrl: 'polyline.html',
-            styleUrls: ['polyline.css']
-        }),
-        __param(3, core_1.Optional()),
-        __param(4, core_1.Optional()), 
-        __metadata('design:paramtypes', [map_service_1.MapService, group_service_1.GroupService, popup_service_1.PopupService, map_1.LeafletElement, group_1.LeafletGroup])
-    ], PolylineElement);
+    PolylineElement.prototype.ngDoCheck = function () {
+        var map = this.mapService.getMap();
+        var same = this.helperService.arrayCompare(this.originalObject, this.latlngs);
+        if (!same) {
+            this.originalObject = this.latlngs.slice();
+            this.Options.fill = false;
+            var inheritedOptions = new path_1.path(this.Options);
+            if (this.LeafletGroup) {
+                this.polyline = L.polyline(this.latlngs, inheritedOptions);
+                this.groupService.addOLayersToGroup(this.polyline, map, this.mapService, this.LeafletGroup, true, this.globalId);
+            }
+            else {
+                map.removeLayer(this.polyline);
+                this.polyline = L.polyline(this.latlngs, inheritedOptions);
+                this.polyline.addTo(map);
+            }
+        }
+    };
     return PolylineElement;
 }());
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Array)
+], PolylineElement.prototype, "latlngs", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Object)
+], PolylineElement.prototype, "Options", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", String)
+], PolylineElement.prototype, "mouseover", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", String)
+], PolylineElement.prototype, "onclick", void 0);
+PolylineElement = __decorate([
+    core_1.Component({
+        moduleId: module.id,
+        selector: 'polyline-element',
+        templateUrl: 'polyline.html',
+        styleUrls: ['polyline.css']
+    }),
+    __param(5, core_1.Optional()),
+    __param(6, core_1.Optional()),
+    __metadata("design:paramtypes", [map_service_1.MapService,
+        group_service_1.GroupService,
+        popup_service_1.PopupService,
+        globalId_service_1.GuidService,
+        helper_service_1.HelperService,
+        map_1.LeafletElement,
+        group_1.LeafletGroup])
+], PolylineElement);
 exports.PolylineElement = PolylineElement;
 //# sourceMappingURL=polyline.js.map
