@@ -11,11 +11,11 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 var __param = (this && this.__param) || function (paramIndex, decorator) {
     return function (target, key) { decorator(target, key, paramIndex); }
 };
-var core_1 = require('@angular/core');
-var map_1 = require('../map/map');
-var group_1 = require('../group/group');
-var map_service_1 = require('../services/map.service');
-var group_service_1 = require('../services/group.service');
+var core_1 = require("@angular/core");
+var map_1 = require("../map/map");
+var group_1 = require("../group/group");
+var map_service_1 = require("../services/map.service");
+var group_service_1 = require("../services/group.service");
 var Lealflet = require('leaflet');
 var PopupElement = (function () {
     function PopupElement(mapService, groupService, LeafletElement, LeafletGroup) {
@@ -30,9 +30,9 @@ var PopupElement = (function () {
     PopupElement.prototype.ngOnInit = function () {
         if (this.LeafletElement || this.LeafletGroup) {
             var map = this.mapService.getMap();
-            var popup = L.popup({ autoClose: false, keepInView: true, closeButton: false }).setLatLng([this.lat, this.lon]).setContent(this.content);
+            var popup = L.popup({ autoClose: false, keepInView: true }).setLatLng([this.lat, this.lon]).setContent(this.content);
             if (this.LeafletGroup) {
-                this.groupService.addOLayersToGroup(popup);
+                this.groupService.addOLayersToGroup(popup, map, this.mapService, this.LeafletGroup);
                 this.groupService.increaseNumber();
             }
             else {
@@ -43,30 +43,33 @@ var PopupElement = (function () {
             console.warn("This popup-element will not be rendered \n the expected parent node of popup-element should be either leaf-element or leaflet-group");
         }
     };
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], PopupElement.prototype, "lat", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], PopupElement.prototype, "lon", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', String)
-    ], PopupElement.prototype, "content", void 0);
-    PopupElement = __decorate([
-        core_1.Component({
-            moduleId: module.id,
-            selector: 'popup-element',
-            templateUrl: 'popup.html',
-            styleUrls: ['popup.css']
-        }),
-        __param(2, core_1.Optional()),
-        __param(3, core_1.Optional()), 
-        __metadata('design:paramtypes', [map_service_1.MapService, group_service_1.GroupService, map_1.LeafletElement, group_1.LeafletGroup])
-    ], PopupElement);
     return PopupElement;
 }());
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], PopupElement.prototype, "lat", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], PopupElement.prototype, "lon", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", String)
+], PopupElement.prototype, "content", void 0);
+PopupElement = __decorate([
+    core_1.Component({
+        moduleId: module.id,
+        selector: 'popup-element',
+        templateUrl: 'popup.html',
+        styleUrls: ['popup.css']
+    }),
+    __param(2, core_1.Optional()),
+    __param(3, core_1.Optional()),
+    __metadata("design:paramtypes", [map_service_1.MapService,
+        group_service_1.GroupService,
+        map_1.LeafletElement,
+        group_1.LeafletGroup])
+], PopupElement);
 exports.PopupElement = PopupElement;
 //# sourceMappingURL=popup.js.map
